pusheo para que Angel no pueda decir que no esta pusheado por la mañana y que como no le respondi por estar dormido no acabo trabajando

+++ Assets/Scripts/Behaviour/HitSpinner.cs
@@ -16,7 +16,7 @@ public class HitSpinner : MonoBehaviour
             transform.eulerAngles += Vector3.up * 5;
         };
 
-        _health.currentHealth = float.MaxValue;
+        _health.currentHealth = int.MaxValue;
     }
 
 }
